+++ src/backends/rustler/gen_bindings/mod.rs
@@ -4,6 +4,8 @@ mod native;
 mod public_api;
 mod public_api_args;
 mod public_api_delegates;
+mod public_api_opaque_methods;
+mod public_api_render;
 mod public_files;
 mod rust_items;
 mod service_api;